+++ Libs/DeathNotificationLib/DeathNotificationLib~Init.lua
@@ -2,7 +2,7 @@
 DeathNotificationLib~Init.lua
 --]]
 
-local VERSION = 21
+local VERSION = 22
 
 if DeathNotificationLib and (DeathNotificationLib.VERSION or 0) >= VERSION then return end